Refactor and initing
